ventas/requirements.txt
kafka-python>=2.0.2
psycopg2-binary>=2.9.1
schedule>=1.1.0
asyncio>=3.4.3